nvim: add zig, refactor descriptions
The descriptions for the keymaps are much more searchable now. For the
ones that constitute some sort of logical group, I followed the same
pattern as the dap bindings, which is just really smooth when using pick
keymaps. Oh, and I also added my custom pickers to the mini.pick
registry.